add: postgres db and entitity/repository
